Refactor unit spawning
There was a race condition between the units, which spawned in Start, and the unit managers, which declared all units to be ready in Start. The unit managers now use the UnitManagerReadyCommands to tell the units when to spawn. This also allows the unit managers to clear their unit lists, which fixes a bug with units from the previous encounter persisting into the next.