Testing out a new ability weighting feature
Ability speed sucks, so we're trying out another approach. Turn queue is randomised but Speed is used as a weight. The ability speed determines the unit speed at the start of the game. So having lots of abilities with low speed means the character is carrying "Heavy" abilities, which makes them more likely to be last in their turns.